Added environments with ball functionality.
Each individual is now evaluated over a number of environments. Each environment contains a ball starting a different position.

Force applied to each ball in the -y direction at the start of each sim.

Collision detection implemented.

Fitness is now defined as the sum of the fitnesses over all the environments divided by the number of environments.